+++ test/unit/mongo_rep_admin/call_func.py
@@ -28,50 +28,44 @@ import version                                  # pylint:disable=E0401,C0413
 __version__ = version.__version__
 
 
-def fetch_priority(repset, args_array, **kwargs):
+def fetch_priority(repset, dtg, **kwargs):
 
     """Function:  fetch_priority
 
     Description:  Stub holder for mongo_rep_admin.fetch_priority function.
 
     Arguments:
-        (input) repset -> Replication set instance.
-        (input) args_array -> Array of command line options and values.
 
     """
 
     err_msg = "Error Message"
-    mail = kwargs.get("mail", None)
     status = (False, err_msg)
 
-    if args_array and repset and mail:
+    if repset and dtg and kwargs.get("suppress", False):
         status = (False, err_msg)
 
     return status
 
 
-def prt_rep_stat(repset, args_array, **kwargs):
+def chk_rep_stat(repset, dtg, **kwargs):
 
-    """Function:  prt_rep_stat
+    """Function:  chk_rep_stat
 
-    Description:  Stub holder for mongo_rep_admin.prt_rep_stat function.
+    Description:  Stub holder for mongo_rep_admin.chk_rep_stat function.
 
     Arguments:
-        (input) repset -> Replication set instance.
-        (input) args_array -> Array of command line options and values.
 
     """
 
-    mail = kwargs.get("mail", None)
     status = (True, None)
 
-    if args_array and repset and mail:
+    if repset and dtg and kwargs.get("suppress", False):
         status = (True, None)
 
     return status
 
 
-class ArgParser():
+class ArgParser():                                      # pylint:disable=R0903
 
     """Class:  ArgParser
 
@@ -80,7 +74,6 @@ class ArgParser():
     Methods:
         __init__
         get_args_keys
-        get_val
 
     """
 
@@ -108,83 +101,6 @@ class ArgParser():
         """
 
         return list(self.args_array.keys())
-
-    def get_val(self, skey, def_val=None):
-
-        """Method:  get_val
-
-        Description:  Method stub holder for gen_class.ArgParser.get_val.
-
-        Arguments:
-
-        """
-
-        return self.args_array.get(skey, def_val)
-
-
-class Coll():
-
-    """Class:  Coll
-
-    Description:  Class stub holder for mongo_class.Coll class.
-
-    Methods:
-        __init__
-        connect
-        coll_cnt
-        coll_find1
-
-    """
-
-    def __init__(self):
-
-        """Method:  __init__
-
-        Description:  Class initialization.
-
-        Arguments:
-
-        """
-
-        self.count = 1
-        self.conn = True
-        self.errmsg = None
-
-    def connect(self):
-
-        """Method:  connect
-
-        Description:  Stub holder for mongo_class.Coll.connect method.
-
-        Arguments:
-
-        """
-
-        return self.conn, self.errmsg
-
-    def coll_cnt(self):
-
-        """Method:  coll_cnt
-
-        Description:  Stub holder for mongo_class.Coll.coll_cnt method.
-
-        Arguments:
-
-        """
-
-        return self.count
-
-    def coll_find1(self):
-
-        """Method:  coll_find1
-
-        Description:  Stub holder for mongo_class.Coll.coll_find1 method.
-
-        Arguments:
-
-        """
-
-        return {"_id": "RepSetName"}
 
 
 class RepSet():                                         # pylint:disable=R0903
@@ -233,7 +149,6 @@ class UnitTest(unittest.TestCase):
         setUp
         test_func_no_error
         test_func_error
-        test_email
 
     """
 
@@ -256,9 +171,11 @@ class UnitTest(unittest.TestCase):
         self.args2.args_array = {
             "-T": True, "-c": "config", "-d": "dirpath", "-e": "Email_Address"}
         self.args3.args_array = {"-P": True, "-c": "config", "-d": "dirpath"}
-        self.func_names = {"-P": fetch_priority, "-T": prt_rep_stat}
+        self.func_names = {"-P": fetch_priority, "-T": chk_rep_stat}
+        self.data_config = {"to_addr": "ToAddress", "suppress": True}
 
-    def test_func_no_error(self):
+    @mock.patch("mongo_rep_admin.create_data_config")
+    def test_func_no_error(self, mock_config):
 
         """Function:  test_func_no_error
 
@@ -268,11 +185,14 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        mock_config.return_value = self.data_config
+
         self.assertFalse(
             mongo_rep_admin.call_func(
                 self.args, self.func_names, self.repset))
 
-    def test_func_error(self):
+    @mock.patch("mongo_rep_admin.create_data_config")
+    def test_func_error(self, mock_config):
 
         """Function:  test_func_error
 
@@ -282,27 +202,12 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        mock_config.return_value = self.data_config
+
         with gen_libs.no_std_out():
             self.assertFalse(
                 mongo_rep_admin.call_func(
                     self.args3, self.func_names, self.repset))
-
-    @mock.patch("mongo_rep_admin.gen_class.setup_mail")
-    def test_email(self, mock_mail):
-
-        """Function:  test_email
-
-        Description:  Test with setting up email.
-
-        Arguments:
-
-        """
-
-        mock_mail.return_value = "Mail Instance"
-
-        self.assertFalse(
-            mongo_rep_admin.call_func(
-                self.args2, self.func_names, self.repset))
 
 
 if __name__ == "__main__":