+++ test/unit/sonarqube_code_coverage.sh
@@ -9,7 +9,8 @@ coverage erase
 echo ""
 echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/call_func.py
-coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/process_std.py
+coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/create_data_config.py
+coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/create_header.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/chk_mem_rep_lag.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/chk_rep_lag.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/chk_rep_stat.py
@@ -20,8 +21,6 @@ coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/get_optimedat
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/help_message.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/main.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/node_chk.py
-coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/process_json.py
-coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/prt_rep_stat.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/rep_health_chk.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/rep_msg_chk.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/rep_state_chk.py